Move to single Activity architecture
