+++ app/src/main/java/com/example/appName/MainActivity.kt
@@ -1,0 +1,42 @@
+package com.example.appName
+
+import android.os.Bundle
+import android.view.MenuItem
+import android.view.WindowManager
+import androidx.appcompat.app.AppCompatActivity
+import com.example.appName.presentation.features.main.MainFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean =
+            when (item.itemId) {
+                android.R.id.home -> {
+                    onBackPressed()
+                    true
+                }
+                else -> super.onOptionsItemSelected(item)
+            }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.activity_main)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager
+                    .beginTransaction()
+                    .add(R.id.main_frame, MainFragment())
+                    .commit()
+        }
+
+        if (!BuildConfig.DEBUG) {
+            window.setFlags(
+                    WindowManager.LayoutParams.FLAG_SECURE,
+                    WindowManager.LayoutParams.FLAG_SECURE
+            )
+        }
+    }
+
+}
