+++ app/src/main/java/com/example/appName/presentation/features/main/MainFragment.kt
@@ -1,24 +1,25 @@
 package com.example.appName.presentation.features.main
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import androidx.activity.viewModels
+import androidx.fragment.app.viewModels
 import com.example.appName.R
-import com.example.appName.presentation.features.base.BaseActivity
+import com.example.appName.presentation.features.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.fragment_main.*
 
 @AndroidEntryPoint
-class MainActivity : BaseActivity<MainViewState, MainPresenter>(
-        R.layout.activity_main
-) {
+class MainFragment : BaseFragment<MainViewState, MainPresenter>(R.layout.fragment_main) {
+
     override val presenter by viewModels<MainPresenter>()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
         login_button.setOnClickListener { presenter.acceptIntent(MainIntent.Login) }
-
         logout_button.setOnClickListener { presenter.acceptIntent(MainIntent.Logout) }
     }
 
@@ -41,5 +42,4 @@ class MainActivity : BaseActivity<MainViewState, MainPresenter>(
             logout_button.visibility = View.GONE
         }
     }
-    //endregion
 }
