+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="action_login">Login</string>
     <string name="action_logout">Logout</string>
     <string name="main_welcome_text">Welcome %s!</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
